create a common file for wonderland tracking things.
